Optimize observation spaces
Actually revert back to using Box for scores because of the calculations speed. The same model with Discrete instead of Box for each score achieves merely about 300 FPS as opposed to > 1000 for Box. It is probably due to the high dimensionality and need to reallocate like 60kB at each step for each environment.